+++ Public/category-map.ps1
@@ -121,6 +121,7 @@ $categoryMap = @{
     'Remove-sqmDatabaseFromAG'            = 'Always On & Availability Groups'
     'Register-sqmBackupExcludeTrigger'    = 'Backup & Recovery'
     'Repair-sqmAlwaysOnDatabases'         = 'Always On & Availability Groups'
+    'Repair-sqmServerName'                = 'Configuration Management'
     'Set-sqmBackupExcludePermission'      = 'Backup & Recovery'
     'Show-sqmBackupExcludeForm'           = 'Backup & Recovery'
     'Set-sqmConfig'                       = 'Configuration Management'